Moved translate_text function to a seperate class - Modified main - Added translation.py
